+++ pulsar/apps/rpc/mixins.py
@@ -1,4 +1,5 @@
-from pulsar import send, coroutine_return
+import pulsar
+from pulsar import coroutine_return
 
 from .jsonrpc import JSONRPC
 
@@ -24,7 +25,7 @@ class PulsarServerCommands(JSONRPC):
         It invokes the :meth:`extra_server_info` for adding custom
         information.
         '''
-        info = yield send('arbiter', 'info')
+        info = yield pulsar.send('arbiter', 'info')
         info = yield self.extra_server_info(request, info)
         coroutine_return(info)
 
@@ -39,7 +40,7 @@ class PulsarServerCommands(JSONRPC):
 
     def rpc_kill_actor(self, request, aid):
         '''Kill the actor with id equal to *aid*.'''
-        return send('arbiter', 'kill_actor', aid)
+        return pulsar.send('arbiter', 'kill_actor', aid)
 
     def extra_server_info(self, request, info):
         '''An internal method.