+++ tests/http/base.py
@@ -425,10 +425,10 @@ class TestHttpClient(TestHttpClientBase, unittest.TestCase):
     def test_send_cookie(self):
         http = self._client
         cookies = {'sessionid': 't1', 'cookies_are': 'working'}
-        response = yield http.get(self.httpbin(), cookies=cookies)
+        response = yield http.get(self.httpbin('cookies'), cookies=cookies)
         self.assertEqual(response.status_code, 200)
-        self.assertTrue(response.headers['set-cookie'])
-        self.assertEqual(response.cookies, SimpleCookie(cookies))
+        data = response.decode_content()
+        self.assertEqual(data['cookies'], cookies)
 
     def test_cookie(self):
         http = self._client